bookstore-api/requirements.txt
Django==2.2.7
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
django-filter==2.2.0
